refactor(web): ip comments folder moved into representations
