--- appeals/web/src/server/appeals/appeal-details/interested-party-comments/common/index.js
@@ -1,5 +1,0 @@
-export { renderDateSubmittedFactory, postDateSubmittedFactory } from '../common/date-submitted.js';
-export {
-	renderRedactionStatusFactory,
-	postRedactionStatusFactory
-} from '../common/redaction-status.js';
