+++ appeals/web/src/server/appeals/appeal-details/representations/interested-party-comments/common/date-submitted.js
@@ -1,54 +1,41 @@
 import { appealShortReference } from '#lib/appeals-formatter.js';
-import { radiosInput } from '#lib/mappers/index.js';
-import { APPEAL_REDACTED_STATUS } from 'pins-data-model';
+import { dateInput } from '#lib/mappers/index.js';
 
-/** @typedef {import("../../appeal-details.types.js").WebAppeal} Appeal */
+/** @typedef {import("../../../appeal-details.types.js").WebAppeal} Appeal */
 /** @typedef {import('#appeals/appeal-details/representations/types.js').Representation} Representation */
-/** @typedef {{ 'redactionStatus': string }} ReqBody */
+/** @typedef {{ 'day': string, 'month': string, 'year': string }} RequestDate */
+/** @typedef {RequestDate} ReqBody */
 
-export const statusFormatMap = {
-	[APPEAL_REDACTED_STATUS.REDACTED]: 'Redacted',
-	[APPEAL_REDACTED_STATUS.NOT_REDACTED]: 'Unredacted',
-	[APPEAL_REDACTED_STATUS.NO_REDACTION_REQUIRED]: 'No redaction required'
-};
-
-export const name = 'redactionStatus';
-
-/**
- * @param {any} maybeRedactionStatus
- * @returns {maybeRedactionStatus is keyof APPEAL_REDACTED_STATUS}
- */
-export const isValidRedactionStatus = (maybeRedactionStatus) =>
-	Object.keys(statusFormatMap).includes(maybeRedactionStatus);
 /**
  * @param {Appeal} appealDetails
  * @param {import('@pins/express').ValidationErrors | undefined} errors
- * @param {string} value
+ * @param {ReqBody} date
  * @param {string} backLinkUrl
  * @returns {PageContent}
- */
-const mapper = (appealDetails, errors, value, backLinkUrl) => ({
-	title: 'Redaction status',
+ * */
+export const mapper = (appealDetails, errors, date, backLinkUrl) => ({
+	title: 'When did the interested party submit the comment?',
 	backLinkUrl,
 	preHeading: `Appeal ${appealShortReference(appealDetails.appealReference)}`,
 	pageComponents: [
-		radiosInput({
-			name,
-			legendText: 'Redaction status',
+		dateInput({
+			id: 'date',
+			name: 'date',
+			value: date,
+			legendText: 'When did the interested party submit the comment?',
 			legendIsPageHeading: true,
-			items: Object.entries(statusFormatMap).map(([value, text]) => ({ value, text })),
-			value
+			hint: 'For example, 27 3 2024'
 		})
 	]
 });
 
 /**
  * @param {object} options
- * @param {(appealDetails: Appeal, comment: Representation) => string} options.getBackLinkUrl,
- * @param {(request: import('@pins/express').Request) => string} options.getValue
+ * @param {(appealDetails: Appeal, comment: Representation) => string} options.getBackLinkUrl
+ * @param {(request: import('@pins/express').Request) => RequestDate} options.getValue
  * @returns {import('@pins/express').RenderHandler<{}, {}, ReqBody>}
  */
-export const renderRedactionStatusFactory =
+export const renderDateSubmittedFactory =
 	({ getBackLinkUrl, getValue }) =>
 	(request, response) => {
 		const backLinkUrl = getBackLinkUrl(request.currentAppeal, request.currentRepresentation);
@@ -68,7 +55,7 @@ export const renderRedactionStatusFactory =
  * @param {import('@pins/express').RenderHandler<{}, {}, ReqBody>} options.errorHandler
  * @returns {import('@pins/express').RenderHandler<{}, {}, ReqBody>}
  */
-export const postRedactionStatusFactory =
+export const postDateSubmittedFactory =
 	({ getRedirectUrl, errorHandler }) =>
 	(request, response, next) => {
 		if (request.errors) {
