+++ appeals/web/src/server/appeals/appeal-details/representations/interested-party-comments/common/index.js
@@ -1,0 +1,2 @@
+export { renderDateSubmittedFactory, postDateSubmittedFactory } from './date-submitted.js';
+export { renderRedactionStatusFactory, postRedactionStatusFactory } from './redaction-status.js';
