+++ appeals/web/src/server/appeals/appeal-details/representations/interested-party-comments/interested-party-comments.router.js
@@ -1,6 +1,6 @@
 import { asyncHandler } from '@pins/express';
 import { Router as createRouter } from 'express';
-import { validateAppeal } from '../appeal-details.middleware.js';
+import { validateAppeal } from '../../appeal-details.middleware.js';
 import addIpCommentRouter from './add-ip-comment/add-ip-comment.router.js';
 import editIpCommentRouter from './edit-ip-comment/edit-ip-comment.router.js';
 import addDocumentRouter from './add-document/add-document.router.js';
