+++ appeals/web/src/server/appeals/appeal-details/representations/interested-party-comments/redact/redact.router.js
@@ -1,6 +1,6 @@
 import { Router as createRouter } from 'express';
 import { asyncHandler } from '@pins/express';
-import { validateAppeal } from '../../appeal-details.middleware.js';
+import { validateAppeal } from '../../../appeal-details.middleware.js';
 import { validateComment } from '../interested-party-comments.middleware.js';
 import {
 	postConfirmRedactInterestedPartyComment,
