+++ appeals/web/src/server/appeals/appeal-details/representations/interested-party-comments/view-and-review/view-and-review.controller.js
@@ -17,7 +17,7 @@ import {
 } from '#appeals/appeal-documents/appeal-documents.controller.js';
 import { render } from '#appeals/appeal-details/representations/common/render.js';
 
-/** @typedef {import("../../appeal-details.types.js").WebAppeal} Appeal */
+/** @typedef {import("../../../appeal-details.types.js").WebAppeal} Appeal */
 /** @typedef {import('#appeals/appeal-details/representations/types.js').Representation} Representation */
 
 export const renderViewInterestedPartyComment = render(
